refactor: 전단지 상세페이지 로직 분리
